ENH: Added new Module RegistrationCuda. Only added the files yet
git-svn-id: http://svn.slicer.org/Slicer4@6171 3bd1e089-480b-0410-8dfb-8563597acbee

+++ branches/igt_cuda_merge/Modules/RegistrationCuda/ntkElasticTransformation.cpp
@@ -147,8 +147,8 @@ ntk3DData* ntkElasticTransformation::getDeformationFieldImage(ntkDeformationSpli
     tempPow=1;
     tempSize=1;
   }else{
-    tempPow=pow(2, splineSizeLevel);
-    tempSize=(int)pow(2, -splineSizeLevel);
+    tempPow=pow((float)2, (float)splineSizeLevel);
+    tempSize=(int)pow((float)2, (float)-splineSizeLevel);
   }
   ntkIntDimension splineSize((int)(tempPow*(m_inputSize.x+1)+1), (int)(tempPow*(m_inputSize.y+1)+1), (int)(tempPow*(m_inputSize.z+1)+1));
