+++ branches/igt_cuda_merge/Modules/RegistrationCuda/ntkElasticTransformationCUDA.cpp
@@ -25,7 +25,7 @@ ntk3DData* ntkElasticTransformationCUDA::applyTransformation(ntkDeformationSplin
   if(splineSizeLevel==0){
     tempPow=1;
   }else{
-    tempPow=pow(2, splineSizeLevel);
+    tempPow=pow((float)2, (float)splineSizeLevel);
   }
 
   ntkIntDimension splineSize((int)(tempPow*(m_inputSize.x+1)+1), (int)(tempPow*(m_inputSize.y+1)+1), (int)(tempPow*(m_inputSize.z+1)+1));
@@ -58,8 +58,8 @@ ntkFloatDimension ntkElasticTransformationCUDA::getNewPositionFromOld(ntkDeforma
     tempPow=1;
     tempSize=1;
   }else{
-    tempPow=pow(2, splineSizeLevel);
-    tempSize=(int)pow(2, -splineSizeLevel);
+    tempPow=pow((float)2,(float) splineSizeLevel);
+    tempSize=(int)pow((float)2, (float)-splineSizeLevel);
   }
   ntkIntDimension splineSize((int)(tempPow*(m_inputSize.x+1)+1), (int)(tempPow*(m_inputSize.y+1)+1), (int)(tempPow*(m_inputSize.z+1)+1));
 
@@ -104,8 +104,8 @@ ntk3DData* ntkElasticTransformationCUDA::getDeformationFieldImage(ntkDeformation
     tempPow=1;
     tempSize=1;
   }else{
-    tempPow=pow(2, splineSizeLevel);
-    tempSize=(int)pow(2, -splineSizeLevel);
+    tempPow=pow((float)2, splineSizeLevel);
+    tempSize=(int)pow((float)2, -splineSizeLevel);
   }
   ntkIntDimension splineSize((int)(tempPow*(m_inputSize.x+1)+1), (int)(tempPow*(m_inputSize.y+1)+1), (int)(tempPow*(m_inputSize.z+1)+1));
